refactor(core): colocate locate prompt spec
