+++ packages/core/tests/unit-test/locate-result-adapter.test.ts
@@ -1,5 +1,5 @@
-import { locateResultExampleRegions } from '@/ai-model/prompts/locate-result-coordinates';
 import { createLocateResultAdapter } from '@/ai-model/shared/model-locate-result';
+import { locateResultExampleRegions } from '@/ai-model/shared/model-locate-result/prompt-spec';
 import { pixelBboxToRect } from '@/ai-model/workflows/inspect/locate-result-rect';
 import { describe, expect, it, vi } from 'vitest';
